+++ NEWS.md
@@ -1,4 +1,4 @@
-# crane 0.3.1.9011
+# crane 0.3.1.9012
 
 ## New Functions and Functionality